requirements.txt
fastapi
uvicorn
azure-cognitiveservices-speech
openai
python-multipart
python-dotenv
ffmpeg-python
